Moved comprehensive test of callable from test_run.py to test_util.py
